- запустил Spring Data JPA - кастомный контроль транзакций перееха в SampleRepository - Examination и Channel пишутся теперь по команде а не гибирнейтом - перееха с @ClassID на @EmbeddedId в классах Channel и Sample - причесал все entity - актуализировал ExaminationServiceTest - рефактеринг

+++ src/main/java/ru/gsa/biointerface/domain/entity/Device.java
@@ -6,7 +6,9 @@ import javax.validation.constraints.Min;
 import javax.validation.constraints.NotNull;
 import javax.validation.constraints.Size;
 import java.io.Serializable;
-import java.util.*;
+import java.util.Objects;
+import java.util.Set;
+import java.util.TreeSet;
 
 /**
  * Created by Gavrilov Stepan (itgavrilov@gmail.com) on 10.09.2021.
@@ -17,7 +19,7 @@ public class Device implements Serializable, Comparable<Device> {
     @NotNull(message = "Id can't be null")
     @Min(value = 1, message = "Id can't be lass then 1")
     @Id
-    private long id;
+    private int id;
 
     @NotNull(message = "Amount channels can't be null")
     @Min(value = 1, message = "Amount channels can't be lass then 1")
@@ -36,18 +38,18 @@ public class Device implements Serializable, Comparable<Device> {
     public Device() {
     }
 
-    public Device(long id, int amountChannels) {
+    public Device(int id, int amountChannels) {
         this.id = id;
         this.amountChannels = amountChannels;
         this.comment = null;
         this.examinations = new TreeSet<>();
     }
 
-    public long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(long id) {
+    public void setId(int id) {
         this.id = id;
     }
