+++ src/main/java/ru/gsa/biointerface/repository/ExaminationRepository.java
@@ -11,6 +11,6 @@ import java.util.List;
  * Created by Gavrilov Stepan (itgavrilov@gmail.com) on 01/11/2021
  */
 @Repository
-public interface ExaminationRepository extends JpaRepository<Examination, Long>{
+public interface ExaminationRepository extends JpaRepository<Examination, Integer> {
     List<Examination> findAllByPatientRecord(PatientRecord patientRecord);
 }
